src/requirements.txt
aniso8601==7.0.0
appdirs==1.4.4
attrs==20.3.0
bcrypt==3.2.0
black==20.8b1
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
click==7.1.2
cryptography==3.4.6
dependency-injector==4.21.0
ecdsa==0.14.1
fastapi==0.63.0
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
h11==0.12.0
idna==2.10
iniconfig==1.1.1
mypy-extensions==0.4.3
packaging==20.9
passlib==1.7.4
pathspec==0.8.1
pluggy==0.13.1
promise==2.3
py==1.10.0
pyasn1==0.4.8
pycparser==2.20
pydantic==1.7.3
pyparsing==2.4.7
pytest==6.2.0
pytest-asyncio==0.14.0
python-jose==3.2.0
regex==2020.11.13
requests==2.25.0
rsa==4.7.2
Rx==1.6.1
six==1.15.0
starlette==0.13.6
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.3
uvicorn==0.13.3
